requirements.txt
Django==2.2.10
django-jsonfield==1.4.1
django-split-json-widget==1.16
djangorestframework==3.12.2
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
